[Feature] change invoke the thread from analyze to discussion component
