+++ components/Analyze.js
@@ -6,21 +6,10 @@ import Category from './CategoryButton';
 import ImageGrid from './ImageGrid';
 import Distributions from './Distributions';
 import Discussion from './Discussion';
-import Thread from './Thread';
 
 import style from '../styles/Analyze.module.css';
-import threadStyle from '../styles/Thread.module.css';
 
 const Analyze = ({ images, distribution, selectedCategory, onSelectCategory }) => {
-
-    const [isThreadVisible, setThreadVisible] = useState(false);
-
-    // click "Start a discussion" button to show the thread
-    const handleStartDiscussion = () => setThreadVisible(true);
-
-    // click '-' to close the thread
-    const handleCloseThread = () => setThreadVisible(false);
-
     return (
         <section className={style.analyzeSection}>
             <div className={style.greyBg}>
@@ -35,13 +24,9 @@ const Analyze = ({ images, distribution, selectedCategory, onSelectCategory }) =
                 </div>
 
                 <div className={style.discussionContainer}>
-                    <Discussion onStartDiscussion={handleStartDiscussion} />
+                    <Discussion />
                 </div>
-
-                {isThreadVisible && (
-                    <Thread onCloseThread={handleCloseThread} />
-                )}
-        </div>
+            </div>
         </section>
     );
 };
